Now flushing after glFenceSync
4.1.2 Signaling
Footnote 3: The simple flushing behavior defined by SYNC_FLUSH_COMMANDS_BIT
will not help when waiting for a fence command issued in another context’s
command stream to complete. Applications which block on a fence sync object
must take additional steps to assure that the context from which the
corresponding fence command was issued has flushed that command to the
graphics pipeline.